+++ Elitism/AG.js
@@ -96,7 +96,7 @@ do
 		console.log("Aptitud del individuo: "+ (parseInt(x)+1) +": "+ individualsFitness[x] +"\n Aptitud relativa: "+ relativeFitness[x]);
 
 	var graphPoints = document.getElementById("Graphic").getAttribute("points");
-	graphPoints += ""+ ((5*generation)+60) +" "+ (500 - (parseInt(bestFitness)*10) +", ");
+	graphPoints += ""+ ((5*generation)+60) +" "+ (500 - (parseFloat(bestFitness)*10) +", ");
 	document.getElementById("Graphic").setAttribute("points", graphPoints);
 	
 	selectedOnes = selection(individuals, relativeFitness);
